move MaxSearch and RefCount into lib.rs
