+++ bucket_map/src/lib.rs
@@ -3,3 +3,6 @@ pub mod bucket_map;
 mod bucket_stats;
 mod bucket_storage;
 mod index_entry;
+
+pub type MaxSearch = u8;
+pub type RefCount = u64;
